@@ -6,53 +6,47 @@
     "dev": "next dev",
     "build": "next build",
     "start": "next start",
-    "lint": "next lint"
+    "lint": "eslint ."
   },
   "dependencies": {
-    "@clerk/nextjs": "^5.1.4",
-    "@clerk/themes": "^1.7.5",
-    "@hookform/resolvers": "^3.1.1",
-    "@netlify/plugin-nextjs": "^4.40.2",
-    "@radix-ui/react-label": "^2.0.2",
-    "@radix-ui/react-menubar": "^1.0.3",
-    "@radix-ui/react-select": "^1.2.2",
-    "@radix-ui/react-slot": "^1.0.2",
-    "@radix-ui/react-tabs": "^1.0.4",
-    "@radix-ui/react-toast": "^1.1.4",
-    "@reduxjs/toolkit": "^1.9.7",
-    "@types/node": "20.4.2",
-    "@types/react": "18.2.15",
-    "@types/react-dom": "18.2.7",
-    "@uploadthing/react": "^5.2.0",
-    "autoprefixer": "10.4.14",
-    "class-variance-authority": "^0.6.1",
-    "clsx": "^1.2.1",
-    "eslint": "8.45.0",
-    "eslint-config-next": "^13.5.4",
-    "eslint-config-prettier": "^8.8.0",
-    "eslint-config-standard": "^17.1.0",
-    "eslint-plugin-tailwindcss": "^3.13.0",
-    "lucide-react": "^0.260.0",
-    "micro": "^10.0.1",
-    "mongoose": "^7.3.4",
-    "next": "^13.5.4",
-    "postcss": "8.4.26",
-    "prettier": "^3.0.0",
-    "react": "^18.2.0",
-    "react-dom": "^18.2.0",
-    "react-hook-form": "^7.45.1",
-    "react-redux": "^8.1.3",
-    "reactjs-popup": "^2.0.6",
-    "redux": "^4.2.1",
-    "svix": "^1.7.0",
-    "tailwind-merge": "^1.13.2",
-    "tailwindcss": "3.3.3",
+    "@clerk/nextjs": "^6.0.0",
+    "@clerk/themes": "^2.2.0",
+    "@hookform/resolvers": "^3.10.0",
+    "@radix-ui/react-label": "^2.1.0",
+    "@radix-ui/react-menubar": "^1.1.0",
+    "@radix-ui/react-select": "^2.1.0",
+    "@radix-ui/react-slot": "^1.1.0",
+    "@radix-ui/react-tabs": "^1.1.0",
+    "@radix-ui/react-toast": "^1.2.0",
+    "@reduxjs/toolkit": "^2.5.0",
+    "@uploadthing/react": "^7.1.0",
+    "class-variance-authority": "^0.7.1",
+    "clsx": "^2.1.0",
+    "lucide-react": "^0.469.0",
+    "mongoose": "^8.9.0",
+    "next": "^16.1.0",
+    "next-themes": "^0.4.6",
+    "radix-ui": "^1.4.3",
+    "react": "^19.0.0",
+    "react-dom": "^19.0.0",
+    "react-hook-form": "^7.71.1",
+    "react-redux": "^9.2.0",
+    "svix": "^1.42.0",
+    "tailwind-merge": "^2.6.0",
     "tailwindcss-animate": "^1.0.6",
-    "typescript": "5.1.6",
-    "uploadthing": "^5.2.0",
-    "zod": "^3.21.4"
+    "uploadthing": "^7.4.0",
+    "zod": "^3.25.76"
   },
   "devDependencies": {
-    "encoding": "^0.1.13"
+    "@eslint/eslintrc": "^3.2.0",
+    "@tailwindcss/postcss": "^4.0.0",
+    "@types/node": "^22.0.0",
+    "@types/react": "^19.0.0",
+    "@types/react-dom": "^19.0.0",
+    "eslint": "^9.0.0",
+    "eslint-config-next": "^16.1.0",
+    "prettier": "^3.4.0",
+    "tailwindcss": "^4.0.0",
+    "typescript": "^5.7.0"
   }
 }
