@@ -1,26 +1,31 @@
 import { ClerkProvider } from "@clerk/nextjs";
-import { Inter } from "next/font/google";
+import { Outfit } from "next/font/google";
+import { ThemeProvider } from "@/components/theme-provider";
 import "../globals.css";
 
 export const metadata = {
-  title: "Next.js",
-  description: "Generated by Next.js",
+  title: "Coral Thread — Auth",
+  description: "Sign in or create your Coral Thread account.",
 };
 
-const inter = Inter({ subsets: ["latin"] });
+const outfit = Outfit({ subsets: ["latin"] });
 
-export default function RootLayout({
+export default function AuthLayout({
   children,
 }: {
   children: React.ReactNode;
 }) {
   return (
     <ClerkProvider>
-      <html lang="en">
-        <body className={`${inter.className} bg-dark-1`}>
-          <main className="flex items-center justify-center min-h-screen">
+      <html lang="en" suppressHydrationWarning>
+        <body className={outfit.className}>
+          <ThemeProvider
+            attribute="class"
+            defaultTheme="dark"
+            themes={["light", "dark"]}
+          >
             {children}
-          </main>
+          </ThemeProvider>
         </body>
       </html>
     </ClerkProvider>
