@@ -23,11 +23,11 @@ async function Page() {
   return (
     <main className="mx-auto flex max-w-3xl flex-col justify-start px-10 py-20">
       <h1 className="head-text">Onboarding</h1>
-      <p className="mt-3 text-base-regular text-light-2">
+      <p className="mt-3 text-base-regular text-muted-foreground">
         Complete your profile now to get started.
       </p>
 
-      <section className="mt-9 bg-dark-2 p-10">
+      <section className="mt-9 bg-card p-10">
         <AccountProfile user={userData} btnTitle="Continue"/>
       </section>
     </main>
