@@ -1,17 +1,20 @@
 import type { Metadata } from "next";
-import { Inter } from "next/font/google";
+import { Outfit } from "next/font/google";
 import { ClerkProvider } from "@clerk/nextjs";
+import { ThemeProvider } from "@/components/theme-provider";
 
 import "../globals.css";
 
 import Topbar from "@/components/shared/Topbar";
 import Bottombar from "@/components/shared/Bottombar";
 import { ReduxProvider } from "@/context/provider";
+import { ThreadDialogProvider } from "@/context/thread-dialog-context";
+import CreateThreadDialog from "@/components/forms/UploadThread";
 
-const inter = Inter({ subsets: ["latin"] });
+const outfit = Outfit({ subsets: ["latin"] });
 
 export const metadata: Metadata = {
-  title: "Threads",
+  title: "Coral Thread",
   description: "Share threads with anyone, anywhere.",
 };
 
@@ -23,17 +26,26 @@ export default function RootLayout({
   return (
     <ClerkProvider>
       <ReduxProvider>
-        <html lang="en">
-          <body className={inter.className}>
-            <Topbar />
+        <html lang="en" suppressHydrationWarning>
+          <body className={outfit.className}>
+            <ThemeProvider
+              attribute="class"
+              defaultTheme="dark"
+              themes={["light", "dark"]}
+            >
+              <ThreadDialogProvider>
+                <Topbar />
 
-            <main className="mt-10">
-              <section className="main-container">
-                <div className="w-full max-w-4xl">{children}</div>
-              </section>
-            </main>
+                <main className="mt-10">
+                  <section className="main-container">
+                    <div className="w-full max-w-4xl">{children}</div>
+                  </section>
+                </main>
 
-            <Bottombar />
+                <Bottombar />
+                <CreateThreadDialog />
+              </ThreadDialogProvider>
+            </ThemeProvider>
           </body>
         </html>
       </ReduxProvider>
