@@ -1,8 +1,8 @@
-import { createNextRouteHandler } from "uploadthing/next";
- 
+import { createRouteHandler } from "uploadthing/next";
+
 import { ourFileRouter } from "./core";
- 
+
 // Export routes for Next App Router
-export const { GET, POST } = createNextRouteHandler({
+export const { GET, POST } = createRouteHandler({
   router: ourFileRouter,
 });
