@@ -1,4 +1,3 @@
-/* eslint-disable camelcase */
 // Resource: https://clerk.com/docs/users/sync-data-to-your-backend
 // Above article shows why we need webhooks i.e., to sync data to our backend
 
@@ -36,7 +35,7 @@ type Event = {
 
 export const POST = async (request: Request) => {
   const payload = await request.json();
-  const header = headers();
+  const header = await headers();
 
   const heads = {
     "svix-id": header.get("svix-id"),
