@@ -27,8 +27,8 @@ const UserCard = ({ id, name, username, imgUrl, personType }: Props) => {
         />
       </div>
       <div className="flex-1 text-ellipsis">
-        <h4 className="text-base-semibold text-light-1">{name}</h4>
-        <p className="text-small-medium text-gray-1">@{username}</p>
+        <h4 className="text-base-semibold text-foreground">{name}</h4>
+        <p className="text-small-medium text-muted-foreground">@{username}</p>
       </div>
 
       <Button
