@@ -1,11 +1,9 @@
 "use client";
 
-import { useState } from "react";
-import UploadThread from "./UploadThread";
 import Image from "next/image";
+import { useThreadDialog } from "@/context/thread-dialog-context";
 
 function PostThread({
-  userId,
   currentUserImg,
   hasImage,
   children,
@@ -15,38 +13,25 @@ function PostThread({
   hasImage: boolean;
   children: React.ReactNode;
 }) {
-  const [isOpen, setIsOpen] = useState(false);
-
-  const parsedUserId = JSON.parse(userId);
-
-  const closeModal = () => setIsOpen(false);
-  const openModal = () => setIsOpen(true);
+  const { open } = useThreadDialog();
 
   return (
     <div className="flex items-center gap-3 w-full py-7">
-      {
-        hasImage && (
-          <Image
-            src={currentUserImg}
-            alt="Profile image"
-            width={48}
-            height={48}
-            className="rounded-full object-cover"
-          />
-        )
-      }
+      {hasImage && (
+        <Image
+          src={currentUserImg}
+          alt="Profile image"
+          width={48}
+          height={48}
+          className="rounded-full object-cover"
+        />
+      )}
       <button
-        onClick={openModal}
-        className="text-light-4 outline-none bg-transparent border-none cursor-pointer"
+        onClick={open}
+        className="text-muted-foreground outline-none bg-transparent border-none cursor-pointer text-left"
       >
         {children}
       </button>
-      <UploadThread
-        userId={parsedUserId}
-        currentUserImg={currentUserImg}
-        isOpen={isOpen}
-        closeModal={closeModal}
-      />
     </div>
   );
 }
