@@ -1,4 +1,8 @@
 import Image from "next/image";
+import Link from "next/link";
+import { Button } from "../ui/button";
+import FollowButton from "./FollowButton";
+import { Separator } from "../ui/separator";
 
 interface Props {
   accountId: string;
@@ -7,7 +11,11 @@ interface Props {
   username: string;
   imgUrl: string;
   bio: string;
-  type?: 'User' | 'Community';
+  type?: "User" | "Community";
+  followersCount?: number;
+  followingCount?: number;
+  threadsCount?: number;
+  isFollowing?: boolean;
 }
 
 const ProfileHeader = ({
@@ -18,7 +26,13 @@ const ProfileHeader = ({
   imgUrl,
   bio,
   type,
+  followersCount = 0,
+  followingCount = 0,
+  threadsCount = 0,
+  isFollowing = false,
 }: Props) => {
+  const isOwnProfile = accountId === authUserId;
+
   return (
     <div className="flex w-full flex-col justify-start">
       <div className="flex items-center justify-between">
@@ -32,17 +46,42 @@ const ProfileHeader = ({
             />
           </div>
           <div className="flex-1">
-            <h2 className="text-left text-heading3-bold text-dark-1">
+            <h2 className="text-left text-heading3-bold text-foreground">
               {name}
             </h2>
-            <p className="text-base-medium text-gray-1">@{username}</p>
+            <p className="text-base-medium text-muted-foreground">@{username}</p>
           </div>
         </div>
+        {type !== "Community" && (
+          <div>
+            {isOwnProfile ? (
+              <Button variant="outline" size="sm" asChild>
+                <Link href="/profile/edit">Edit Profile</Link>
+              </Button>
+            ) : (
+              <FollowButton
+                currentUserId={authUserId}
+                targetUserId={accountId}
+                isFollowing={isFollowing}
+              />
+            )}
+          </div>
+        )}
       </div>
 
-      <p className="mt-6 max-w-lg text-base-regular text-dark-2">{bio}</p>
+      {bio && (
+        <p className="mt-4 max-w-lg text-base-regular text-foreground">{bio}</p>
+      )}
 
-      <div className="mt-12 h-0.5 w-full bg-neutral-300" />
+      {type !== "Community" && (
+        <div className="mt-4 flex items-center gap-4 text-small-medium text-muted-foreground">
+          <span>{threadsCount} threads</span>
+          <span>{followersCount} followers</span>
+          <span>{followingCount} following</span>
+        </div>
+      )}
+
+      <Separator className="mt-6" />
     </div>
   );
 };
