@@ -2,12 +2,12 @@ function RightSidebar() {
   return (
     <section className="custom-scrollbar rightsidebar">
       <div className="flex flex-1 flex-col justify-start">
-        <h3 className="text-heading4-medium text-light-1">
+        <h3 className="text-heading4-medium text-foreground">
           Suggested Communities
         </h3>
       </div>
       <div className="flex flex-1 flex-col justify-start">
-        <h3 className="text-heading4-medium text-light-1">
+        <h3 className="text-heading4-medium text-foreground">
           Suggested Users
         </h3>
       </div>
