@@ -1,6 +1,6 @@
-import { Middleware } from "redux";
+import type { Middleware } from "@reduxjs/toolkit";
 
-export const threads: Middleware = (storeAPI) => (dispatch) => async (action) => {
+export const threads: Middleware = (storeAPI) => (dispatch) => async (action: any) => {
 
   return dispatch(action);
 
