@@ -1,4 +1,4 @@
-import { generateReactHelpers } from "@uploadthing/react/hooks";
+import { generateReactHelpers } from "@uploadthing/react";
 
 import type { OurFileRouter } from "@/app/api/uploadthing/core";
 
